feat: Intro, Login, Signup の画面 UI 作成
ロゴや OR SIGN IN WITH は仕様揺れてそうなので後回しにした.
細かい箇所も後で Figma に合わせる必要がある.

/login, /signup は モーダルで表現しており, 現在のLoginPageは実質
/intro なので, クラス名も LoginPage→IntroPageに変更.
