+++ .vscode/settings.json
@@ -5,6 +5,7 @@
         "autovalidate",
         "commentq",
         "Kamishibai",
+        "LTRB",
         "signup",
         "thankyouf",
         "thankyouq",
